+++ ibm/service/drautomationservice/resource_ibm_pdr_validate_apikey.go
@@ -19,8 +19,7 @@ import (
 	"github.com/IBM-Cloud/terraform-provider-ibm/ibm/flex"
 	"github.com/IBM/go-sdk-core/v5/core"
 
-	// "github.com/IBM/dra-go-sdk/drautomationservicev1"
-	"github.ibm.com/DRAutomation/dra-go-sdk/drautomationservicev1"
+	"github.com/IBM/dra-go-sdk/drautomationservicev1"
 )
 
 func ResourceIBMPdrValidateApikey() *schema.Resource {